requirements.txt
fastapi
uvicorn[standard]
spotipy
ytmusicapi
python-dotenv
jinja2
